support for private repos; include negnu

+++ setup.py
@@ -44,15 +44,77 @@ def find_include(branch):
     return list(getattr(foo, 'include', ['include']))
 
 
+def download(url, target, authorization=None, size=None):
+    """
+    Download file from input ``url``.
+
+    Parameters
+    ----------
+    url : str, Path
+        url to download file from.
+
+    target : str, Path
+        Path where to save the file, on disk.
+
+    size : int, default=None
+        Expected file size, in bytes, used to show progression bar.
+        If not provided, taken from header (if the file is larger than a couple of GBs,
+        it may be wrong due to integer overflow).
+        If a sensible file size is obtained, a progression bar is printed.
+    """
+    # Adapted from https://stackoverflow.com/questions/15644964/python-progress-bar-and-downloads
+    print('Downloading {} to {}.'.format(url, target))
+    import requests
+    # See https://stackoverflow.com/questions/61991164/python-requests-missing-content-length-response
+    headers = {}
+    if authorization:
+        headers.update({'Authorization': authorization})
+    if size is None:
+        size = requests.head(url, headers={**headers, 'Accept-Encoding': None}).headers.get('content-length')
+    try:
+        r = requests.get(url, headers=headers, allow_redirects=True, stream=True)
+        r.raise_for_status()
+    except requests.exceptions.HTTPError:
+        return False
+
+    with open(target, 'wb') as file:
+        if size is None or int(size) < 0:  # no content length header
+            file.write(r.content)
+        else:
+            import shutil
+            width = shutil.get_terminal_size((80, 20))[0] - 9  # pass fallback
+            dl, size, current = 0, int(size), 0
+            for data in r.iter_content(chunk_size=2048):
+                dl += len(data)
+                file.write(data)
+                if size:
+                    frac = min(dl / size, 1.)
+                    done = int(width * frac)
+                    if done > current:  # it seems, when content-length is not set iter_content does not care about chunk_size
+                        print('\r[{}{}] [{:3.0%}]'.format('#' * done, ' ' * (width - done), frac), end='', flush=True)
+                        current = done
+            print('')
+    return True
+
+
 def build_class(build_dir, branch='base'):
     """Function to dowwnload CLASS from github and build the library."""
     # latest class version and download link
     url = find_url(branch)
+    authorization = os.getenv('AUTHORIZATION', None)
+    depends_dir = os.path.join(package_basedir, 'depends')
+    tarball = 'tmp-class-{}.tar.gz'.format(branch)
+    if not download(url, target=os.path.join(depends_dir, tarball), authorization=authorization):
+        print('\033[93mCould not access {}; skipping branch {}.\033[0m'.format(url, branch))
+        #import warnings
+        #warnings.warn('Could not access {}; skipping branch {}.'.format(url, branch))
+        return False
     patch = os.path.join(os.path.join(package_basedir, package_basename, branch, 'patch'))
-    args = (package_basedir, url, patch, os.path.abspath(build_dir), ' '.join(find_include(branch)))
-    command = 'cd {}/depends; URL={} PATCH={} DEST={} INCLUDE="{}" make install'.format(*args)
+    args = (depends_dir, tarball, patch, os.path.abspath(build_dir), ' '.join(find_include(branch)))
+    command = 'cd {}; TARBALL="{}" PATCH={} DEST={} INCLUDE="{}" make install'.format(*args)
     if os.system(command) != 0:
         raise ValueError('could not build CLASS {}'.format(build_dir))
+    return True
 
 
 class custom_build_ext(build_ext):
@@ -65,25 +127,30 @@ class custom_build_ext(build_ext):
         self.cython_directives = {'language_level': '3' if sys.version_info.major >= 3 else '2'}
 
     def run(self):
+        nobuild = []
         for extension in self.extensions:
             branch = extension.name[len(package_basename) + 1:-len('binding') - 1]
             build_dir = os.path.join(self.build_temp, branch)
-            build_class(build_dir, branch=branch)
-            library_dir = os.path.join(build_dir, 'lib')
-            # os.rename(os.path.join(library_dir, 'libclass.a'), os.path.join(library_dir, 'libclass-{}.a'.format(branch)))
-            for include in find_include(branch):
-                extension.include_dirs.insert(0, os.path.join(build_dir, include))
+            if build_class(build_dir, branch=branch):
+                library_dir = os.path.join(build_dir, 'lib')
+                # os.rename(os.path.join(library_dir, 'libclass.a'), os.path.join(library_dir, 'libclass-{}.a'.format(branch)))
+                for include in find_include(branch):
+                    extension.include_dirs.insert(0, os.path.join(build_dir, include))
 
-            for external in ['heating', 'HyRec2020', 'RecfastCLASS']:
-                extension.include_dirs.insert(0, os.path.join(build_dir, 'external', external))
-            extension.library_dirs.insert(0, library_dir)
-            extension.libraries.insert(0, 'class')
-            # extension.include_dirs = self.include_dirs + extension.include_dirs
+                for external in ['heating', 'HyRec2020', 'RecfastCLASS']:
+                    extension.include_dirs.insert(0, os.path.join(build_dir, 'external', external))
+                extension.library_dirs.insert(0, library_dir)
+                extension.libraries.insert(0, 'class')
+                # extension.include_dirs = self.include_dirs + extension.include_dirs
 
-            # copy data files from temp to pyclass package directory
-            for name in ['external', 'data']:
-                shutil.rmtree(os.path.join(self.build_lib, package_basename, branch, name), ignore_errors=True)
-                shutil.copytree(os.path.join(self.build_temp, branch, name), os.path.join(self.build_lib, package_basename, branch, name))
+                # copy data files from temp to pyclass package directory
+                for name in ['external', 'data']:
+                    shutil.rmtree(os.path.join(self.build_lib, package_basename, branch, name), ignore_errors=True)
+                    shutil.copytree(os.path.join(self.build_temp, branch, name), os.path.join(self.build_lib, package_basename, branch, name))
+            else:
+                nobuild.append(extension)
+        for extension in nobuild:
+            del self.extensions[self.extensions.index(extension)]
 
         # self.include_dirs.clear()
         super(custom_build_ext, self).run()
@@ -97,7 +164,8 @@ class custom_develop(develop):
         for branch in find_branches():
             for name in ['external', 'data']:
                 shutil.rmtree(os.path.join(package_basedir, package_basename, branch, name), ignore_errors=True)
-                shutil.copytree(os.path.join(build_ext.build_temp, branch, name), os.path.join(package_basedir, package_basename, branch, name))
+                try: shutil.copytree(os.path.join(build_ext.build_temp, branch, name), os.path.join(package_basedir, package_basename, branch, name))
+                except OSError: pass
 
 
 class custom_clean(clean):
@@ -108,7 +176,7 @@ class custom_clean(clean):
         super(custom_clean, self).run()
 
         # remove CLASS tmp directories
-        for dirpath in glob.glob(os.path.join('depends', 'tmp*')):
+        for dirpath in glob.glob(os.path.join(package_basedir, 'depends', 'tmp*')):
             if os.path.isdir(dirpath):
                 shutil.rmtree(dirpath)
             else:
